requirements.txt
# Dépendances principales
tqdm>=4.65.0
rich>=13.0.0
psutil>=5.9.0
pywin32>=306; platform_system == "Windows"
pycryptodome>=3.19.0
requests>=2.31.0
yara-python>=4.3.0
volatility3>=2.4.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == "Windows"
pyyaml>=6.0.1
cryptography>=41.0.7
PyJWT>=2.8.0
passlib>=1.7.4
ipaddress>=1.0.23
python-dotenv>=1.0.0
setuptools>=68.0.0
wheel>=0.41.0
pandas>=2.0.0
seaborn>=0.12.0
matplotlib>=3.7.0
reportlab>=4.0.0
bleach>=6.0.0
sqlparse>=0.4.0

# Dépendances de test
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.3.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.1
pytest-timeout>=2.1.0
pytest-env>=1.0.0
pytest-sugar>=0.9.7
pytest-html>=3.2.0
pytest-metadata>=3.0.0
pytest-ordering>=0.6.0
pytest-randomly>=3.12.0
pytest-repeat>=0.9.1
pytest-rerunfailures>=11.1.2
pytest-selenium>=4.0.0
pytest-vcr>=1.0.2
pytest-watch>=4.2.0
pytest-xprocess>=0.22.0
pytest-xvfb>=3.0.0
typing-extensions>=4.9.0
pathlib>=1.0.1
